move settings into a component
