+++ src/components/Settings/index.ts
@@ -1,0 +1,3 @@
+import Settings from './Settings';
+
+export default Settings;
