+++ src/components/WallpaperSource/index.ts
@@ -1,0 +1,3 @@
+import WallpaperSource from './WallpaperSource';
+
+export default WallpaperSource;
